rename graphics.py to battle.py
