+++ battle.py
